1.全部患者随访列表  增加了deptId 科室id  和  deptName 科室名称 2.http://tqmssjyservice.shijiyunyi.com/business/tbFollowVisitPatientTemplate/info 数据修改了

+++ api/src/main/java/com/qu/modules/web/vo/TbFollowVisitPatientTemplateInfoVo.java
@@ -55,7 +55,7 @@ public class TbFollowVisitPatientTemplateInfoVo {
     private Integer followVisitCountNumber;
 
     @ApiModelProperty(value = "已随访次数")
-    private Integer alreadyFollowVisitCountNumber;
+    private Long alreadyFollowVisitCountNumber;
 
     @ApiModelProperty(value = "填报记录(历史随访记录)")
     private List<TbFollowVisitPatientTemplateHistoryVo> historyRecord;
